Refactor out redundant version computation
